[TEST] Move common exception tests in `Exceptions` folder
